fix: P2残Issue6件を一括対応（テスト・認証・認可・リファクタ）
- #24: requireCaseAccess異常系テスト追加（getCase例外→500）
- #32: Firebase Adminエラーコード分岐改善（文字列→code判定）
- #23: admin全件ケース取得（listAllCases追加）
- #26: staffドキュメント削除防止（Firestore Rules delete禁止）
- #22: Firestore Security Rules細粒度化（所有権チェック+Custom Claims対応）
- #19: ルート層の責務分離（cases.ts→consultations.ts分割）

Closes #24, Closes #32, Closes #23, Closes #26, Closes #22, Closes #19
